[PRMT-TECH] ignored permanently vulnerability related to libpq
libpq is a C library to interact with Postgres; the vulnerability flagged has not
been addressed yet.
The risk seems a possible denial of service, if the dependency is invoked wrongly.

We don't interact with libpq directly but via other libraries, and if we were vulnerable
to it we should experience a denial of service already.

This has been ignored for a month previouly and it seems there is no fix in sight,
therefore I ignore this permanently.

For more details:

https://github.com/advisories/GHSA-j32j-2hxv-rqf7
https://nvd.nist.gov/vuln/detail/CVE-2022-25852
https://security.snyk.io/vuln/SNYK-JS-LIBPQ-2392366